Implement battery cam features for Reolink
Allows for a camera to be designated as a battery powered camera, which triggers streaming only when the camera is detecting motion (up to the motion timeout) and reverts to an idle animation when the camera is sleeping. For battery cameras, the snapshot will be cached from the last motion so that it will not show the idle animation.